+++ source/renderer/app/components/settings/categories/TermsOfUseSettings.js
@@ -6,15 +6,19 @@ import styles from './TermsOfUseSettings.scss';
 
 type Props = {
   localizedTermsOfUse: string,
+  onOpenExternalLink: Function,
 };
 
 @observer
 export default class TermsOfUseSettings extends Component<Props> {
   render() {
-    const { localizedTermsOfUse } = this.props;
+    const { localizedTermsOfUse, onOpenExternalLink } = this.props;
     return (
       <div className={styles.component}>
-        <TermsOfUseText localizedTermsOfUse={localizedTermsOfUse} />
+        <TermsOfUseText
+          localizedTermsOfUse={localizedTermsOfUse}
+          onOpenExternalLink={onOpenExternalLink}
+        />
       </div>
     );
   }